refactor(desktop): address PR review on notifications toggle
- Extract shared getNotificationSoundsMuted/setNotificationSoundsMuted into
  main/lib/notification-settings; menu, notification-sound, and the settings
  tRPC router now share one source of truth (removes duplicated read/upsert).
- Wire reverse sync: the shared setter emits notifications-muted-changed, so
  toggling mute from the command palette/settings rebuilds the menu-bar
  checkbox. menu.ts registers the rebuild listener once.
- layout.tsx uses typed utils.settings.getNotificationSoundsMuted.invalidate()
  instead of a hardcoded query key; drop electronQueryClient import.
- Replace misleading rebuild comment in the menu click handler.
