+++ apps/desktop/src/main/lib/menu.ts
@@ -1,4 +1,3 @@
-import { settings } from "@superset/local-db";
 import { COMPANY } from "@superset/shared/constants";
 import { app, BrowserWindow, Menu, shell } from "electron";
 import { env } from "main/env.main";
@@ -9,32 +8,16 @@ import {
 	simulateError,
 	simulateUpdateReady,
 } from "./auto-updater";
-import { localDb } from "./local-db";
 import { menuEmitter } from "./menu-events";
+import {
+	getNotificationSoundsMuted,
+	setNotificationSoundsMuted,
+} from "./notification-settings";
 
-function getNotificationSoundsMuted(): boolean {
-	try {
-		const settingsRow = localDb.select().from(settings).get();
-		return settingsRow?.notificationSoundsMuted ?? false;
-	} catch {
-		return false;
-	}
-}
-
-function setNotificationSoundsMuted(muted: boolean): void {
-	try {
-		localDb
-			.insert(settings)
-			.values({ id: 1, notificationSoundsMuted: muted })
-			.onConflictDoUpdate({
-				target: settings.id,
-				set: { notificationSoundsMuted: muted },
-			})
-			.run();
-	} catch (error) {
-		console.error("[menu] Failed to persist notification mute state:", error);
-	}
-}
+// Registered once (see createApplicationMenu) so the menu-bar checkbox stays in
+// sync when the mute state changes from other surfaces (command palette,
+// settings). Guards against re-registering on every menu rebuild.
+let muteSyncListenerRegistered = false;
 
 export function createApplicationMenu() {
 	const reloadAccelerator = "CmdOrCtrl+R";
@@ -115,11 +98,20 @@ export function createApplicationMenu() {
 					type: "checkbox",
 					checked: getNotificationSoundsMuted(),
 					click: (menuItem) => {
-						setNotificationSoundsMuted(menuItem.checked);
-						menuEmitter.emit("notifications-muted-changed", menuItem.checked);
-						// Rebuild so the checkbox reflects the persisted state if other
-						// surfaces (command palette, settings) change it later.
-						createApplicationMenu();
+						try {
+							// Persisting emits "notifications-muted-changed", which both
+							// rebuilds this menu (via the listener below) and re-syncs the
+							// renderer.
+							setNotificationSoundsMuted(menuItem.checked);
+						} catch (error) {
+							console.error(
+								"[menu] Failed to persist notification mute state:",
+								error,
+							);
+							// Write failed — rebuild to revert the optimistic checkbox toggle
+							// back to the persisted value.
+							createApplicationMenu();
+						}
 					},
 				},
 			],
@@ -233,4 +225,15 @@ export function createApplicationMenu() {
 
 	const menu = Menu.buildFromTemplate(template);
 	Menu.setApplicationMenu(menu);
+
+	// Rebuild the menu whenever the mute state changes — from this menu, the
+	// command palette, or the settings page — so the checkbox reflects the
+	// persisted value. Registered once; the rebuild itself does not emit, so
+	// this can't loop.
+	if (!muteSyncListenerRegistered) {
+		muteSyncListenerRegistered = true;
+		menuEmitter.on("notifications-muted-changed", () => {
+			createApplicationMenu();
+		});
+	}
 }
