+++ apps/desktop/src/renderer/routes/_authenticated/layout.tsx
@@ -21,7 +21,6 @@ import { authClient, getAuthToken } from "renderer/lib/auth-client";
 import { dragDropManager } from "renderer/lib/dnd";
 import { electronTrpc } from "renderer/lib/electron-trpc";
 import { showWorkspaceAutoNameWarningToast } from "renderer/lib/workspaces/showWorkspaceAutoNameWarningToast";
-import { electronQueryClient } from "renderer/providers/ElectronTRPCProvider";
 import { InitGitDialog } from "renderer/react-query/projects/InitGitDialog";
 import { DaemonAutoUpdateFailureDialog } from "renderer/routes/_authenticated/components/DaemonAutoUpdateFailureDialog";
 import { DashboardNewWorkspaceModal } from "renderer/routes/_authenticated/components/DashboardNewWorkspaceModal";
@@ -157,9 +156,7 @@ function AuthenticatedLayout() {
 			} else if (event.type === "open-workspace") {
 				navigate({ to: `/workspace/${event.data.workspaceId}` });
 			} else if (event.type === "notifications-muted-changed") {
-				electronQueryClient.invalidateQueries({
-					queryKey: [["settings", "getNotificationSoundsMuted"]],
-				});
+				utils.settings.getNotificationSoundsMuted.invalidate();
 			}
 		},
 	});
